Add slides.html redirect target
I find myself baking in a long link to a specific slide deck into our
documentation.  This makes it hard to change this link in the future.

I figured it would be safer to link everything to dask.org/slides.html
and then have that page redirect to wherever we like the future.